// enums/promptEnum.ts
export enum PromptEnum {
    ADMIN_WELCOME_MESSAGE_SETUP_PROMPT = 'WELCOME_MESSAGE_SETUP_PROMPT',
    ADMIN_WORKFLOW_DETECTION_PROMPT = 'ADMIN_WORKFLOW_DETECTION_PROMPT',
    ADMIN_CHANNEL_RECOMMENDATIONS_PROMPT = 'ADMIN_CHANNEL_RECOMMENDATIONS_PROMPT',
    ADMIN_SERVER_RULES_PROMPT = 'ADMIN_SERVER_RULES',
}