// enums/workflowEnum.ts
export enum WorkflowEnum {
    ADMIN_WELCOME_MESSAGE = 'welcome_message',
    ADMIN_SERVER_RULES = 'server_rules',
    ADMIN_CHANNEL_RECOMMENDATION = 'channel_recommendation',
    ADMIN_MESSAGE_SEND = 'message_send',
    ADMIN_TOOL_EXECUTE = 'tool_execute',
    USER_TOOL_EXECUTE = 'tool_execute',
    USER_CHANNEL_RECOMMENDATION = 'channel_recommendation',
    UNKNOWN = 'unknown',
}